Split progress refresh tests
