+++ apps/ios/Flashcards/FlashcardsTests/Progress/ProgressDisconnectedFallbackTests.swift
@@ -1,0 +1,67 @@
+import Foundation
+import XCTest
+@testable import Flashcards
+
+final class ProgressDisconnectedFallbackTests: ProgressStoreTestCase {
+    @MainActor
+    func testRefreshProgressIfNeededUsesLocalFallbackWhenDisconnected() async throws {
+        let database = try self.makeDatabase()
+        let workspace = try database.workspaceSettingsStore.loadWorkspace()
+        let cloudSettings = try database.workspaceSettingsStore.loadCloudSettings()
+        try self.addReviewedCard(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            reviewedAtClient: "2026-04-02T15:50:57.000Z"
+        )
+        let secondWorkspace = try self.insertWorkspace(
+            database: database,
+            name: "Workspace 2",
+            createdAt: "2026-04-02T00:00:00Z"
+        )
+        try self.addReviewedCard(
+            database: database,
+            workspaceId: secondWorkspace.workspaceId,
+            reviewedAtClient: "2026-04-03T09:00:00.000Z"
+        )
+
+        let now = try XCTUnwrap(parseIsoTimestamp(value: "2026-04-18T12:00:00.000Z"))
+        let requestRange = try makeTestProgressRequestRange(
+            now: now,
+            timeZone: TimeZone.current,
+            dayCount: 140
+        )
+        let serverSeries = try makeTestProgressSeries(
+            requestRange: requestRange,
+            reviewCountsByDate: [
+                "2026-04-01": 5
+            ],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let serverSummary = try makeTestProgressSummary(
+            timeZone: requestRange.timeZone,
+            reviewDates: ["2026-04-01"],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let context = try self.makeProgressStoreContext(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            installationId: cloudSettings.installationId,
+            serverSummary: serverSummary,
+            serverSeries: serverSeries,
+            loadProgressSummaryError: nil,
+            loadProgressSeriesError: nil,
+            cloudState: .disconnected
+        )
+        defer { context.tearDown() }
+
+        await context.store.refreshProgressIfNeeded(now: now)
+
+        XCTAssertTrue(context.cloudSyncService.recordedOperations.isEmpty)
+        let progressSnapshot = try XCTUnwrap(context.store.progressSnapshot)
+        XCTAssertEqual(.localOnly, progressSnapshot.summarySourceState)
+        XCTAssertEqual(.localOnly, progressSnapshot.seriesSourceState)
+        XCTAssertTrue(progressSnapshot.isApproximate)
+        XCTAssertEqual(2, progressSnapshot.summary.activeReviewDays)
+        XCTAssertEqual("2026-04-03", progressSnapshot.summary.lastReviewedOn)
+    }
+}
